# WIP distsql: Replace Datums in RowFetcher with EncDatums