--- a/pkg/sql/distsql/aggregator_test.go
+++ b/pkg/sql/distsql/aggregator_test.go
@@ -41,7 +41,7 @@ func TestAggregator(t *testing.T) {
 
 	v := [15]sqlbase.EncDatum{}
 	for i := range v {
-		v[i].SetDatum(sqlbase.ColumnType_INT, parser.NewDInt(parser.DInt(i)))
+		v[i] = sqlbase.EncDatumFromDatum(sqlbase.ColumnType_INT, parser.NewDInt(parser.DInt(i)))
 	}
 
 	testCases := []struct {