--- a/pkg/sql/sqlbase/testutils.go
+++ b/pkg/sql/sqlbase/testutils.go
@@ -133,9 +133,7 @@ func RandDatumEncoding(rng *rand.Rand) DatumEncoding {
 func RandEncDatum(rng *rand.Rand) EncDatum {
 	typ := RandColumnType(rng)
 	datum := RandDatum(rng, typ, true)
-	var ed EncDatum
-	ed.SetDatum(typ, datum)
-	return ed
+	return EncDatumFromDatum(typ, datum)
 }
 
 // RandEncDatumSlice generates a slice of random EncDatum values of the same random
@@ -144,7 +142,7 @@ func RandEncDatumSlice(rng *rand.Rand, numVals int) []EncDatum {
 	typ := RandColumnType(rng)
 	vals := make([]EncDatum, numVals)
 	for i := range vals {
-		vals[i].SetDatum(typ, RandDatum(rng, typ, true))
+		vals[i] = EncDatumFromDatum(typ, RandDatum(rng, typ, true))
 	}
 	return vals
 }