… pkg/sql/distsql/distinct_test.go
… pkg/sql/distsql/evaluator_test.go
… pkg/sql/distsql/hashjoiner_test.go
… pkg/sql/distsql/input_sync_test.go
… pkg/sql/distsql/joinreader_test.go
… pkg/sql/distsql/mergejoiner_test.go
… pkg/sql/distsql/sorter_test.go
… pkg/sql/distsql/stream_data_test.go
… pkg/sql/sqlbase/encoded_datum_test.go
… pkg/sql/sqlbase/table.go
… pkg/sql/sqlbase/table_test.go