This is a work in progress, out for preliminary feedback to ensure I'm not 
doing something extremely flawed. 

RowFetcher should return EncDatums. Downstream users who expect a
Datum should extract it from the EncDatum using EncDatumToDTuple.

Other minor changes:
1) Rename EncDatum.Decode to EncDatum.EnsureDecoded.

2) Change SetDatum/SetEncoded/SetFromBuffer from methods that 
mutate structs to functions that return values 

3) (WIP) Merge the functionality of DatumToEncDatum with
SetDatum. Almost done, with the exception of RowFetcher.ProcessKV,
which requires converting RowFetcher.keyVals to []EncDatum.

<!-- Reviewable:start -->

---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/10550)
<!-- Reviewable:end -->