@RaduBerinde 2016-11-08
CC @paperstreet 

Thanks for tackling this. This part looks good so far; the main part that's TBD is to add versions of `DecodeKeyVals`, `UnmarshalColumnValue`, `DecodeTableValue` that use `CreateEncDatumFromBuffer` and return `EncDatum`s without creating any `Datum`s internally.

---

Review status: 0 of 12 files reviewed at latest revision, 4 unresolved discussions.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 67 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4ZY_F1a0nDIEev3v-:-KW4ZY_F1a0nDIEev3v0:bu0qeni) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L67)):*
> ```Go
> // make sure the slice is not modified for the lifetime of the EncDatum.
> // SetEncoded wipes the underlying Datum.
> func CreateEncDatumFromEncoded(typ ColumnType_Kind, enc DatumEncoding, val []byte) EncDatum {
> ```

Thanks for changing these over, this makes more sense. Consider removing the Create, `EncDatumFromXXX` should be clear enough.

Also, in the future try doing cleanup/renamings like this in a separate commit, makes things easier to review and track.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 107 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4augsW655yORqPTI4:-KW4augsW655yORqPTI5:biimxb6) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L107)):*
> ```Go
> 
> // SetDatum initializes the EncDatum with the given Datum.
> func (ed *EncDatum) SetDatum(typ ColumnType_Kind, d parser.Datum) {
> ```

This should also be changed to `CreateEncDatumFromDatum` like the others

---

*[pkg/sql/sqlbase/encoded_datum.go, line 280 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4Z9nQt-3L-ZYVkZnv:-KW4Z9nRMMfxvpfTwMpS:bnlnuak) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L280)):*
> ```Go
> 
> // EncDatumRowToDTuple converts a given EncDatumRow to a DTuple.
> func EncDatumRowToDTuple(tuple parser.DTuple, row EncDatumRow) error {
> ```

This function should be passed a `*DatumAlloc`, and the RowFetcher should have a DatumAlloc member that we use all the time. The point is to amortize the allocations over multiple rows

---

*[pkg/sql/sqlbase/rowfetcher.go, line 510 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4aiQluhJy3b0jzisp:-KW4aiQluhJy3b0jzisq:b-y8p6vo) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/rowfetcher.go#L510)):*
> ```Go
> 		return nil, nil
> 	}
> 	tuple := make(parser.DTuple, len(encRow))
> ```

this row should be allocated once and reused

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-:-KW4ZlUeDyv_y2gohBS8:b-d5qsu)*
<!-- Sent from Reviewable.io -->