@arjunravinarayan 2016-11-26
PTAL @raduberinde and @paperstreet. 

@paperstreet just need your input for the last bit that needs to be done: Changing over `ProcessValueSingle` to no longer use `UnmarshalColumnValue`, which is subtle and tricksy and eluding me. 

@raduberinde I've changed `ProcessValueTuple`, `DecodeKeyVals`, `ProcessKV` to not create intermediate `Datum`s. What remains to be done is `ProcessValueSingle`, and removing `valNeededForCol` from `RowFetcher`. I think the latter is best done in a separate PR (I've marked it as a TODO in the code), but the former should be doable with a little help from @paperstreet.

---

Review status: 0 of 26 files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 67 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4ZY_F1a0nDIEev3v-:-KW9DySubgRzb6W2vOBE:baiu8n) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L67)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> Thanks for changing these over, this makes more sense. Consider removing the Create, `EncDatumFromXXX` should be clear enough.
> 
> Also, in the future try doing cleanup/renamings like this in a separate commit, makes things easier to review and track.
</details>
Removed the `Create` prefixes.

I have also rebased the commit into two separate commits. Apologies for this, in retrospect I'm in a bit of a rebase hell and should have done these as separate PRs.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 107 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4augsW655yORqPTI4:-KW9DVRCaI2LBK5RmlnE:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L107)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> This should also be changed to `CreateEncDatumFromDatum` like the others
</details>
Done.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 280 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4Z9nQt-3L-ZYVkZnv:-KWFB6BEY64Ku08tc8WG:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/encoded_datum.go#L280)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> This function should be passed a `*DatumAlloc`, and the RowFetcher should have a DatumAlloc member that we use all the time. The point is to amortize the allocations over multiple rows
</details>
Done.

---

*[pkg/sql/sqlbase/rowfetcher.go, line 510 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KW4aiQluhJy3b0jzisp:-KWFB7GKyGZKqzQQ-z7-:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/ab34e72b647e591403c8e1a08632299cf0bc5080/pkg/sql/sqlbase/rowfetcher.go#L510)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> this row should be allocated once and reused
</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-:-KXTtd-yI4xL_EIg77lA:bei5k38)*
<!-- Sent from Reviewable.io -->