@arjunravinarayan 2016-11-29
TFTRs! Working on your helpful comments after fixing the column family thing.
Also rebased. Performance benchmarking shows 0%-3% time increase. Working on doing some `count` benchmarks too, which should show time decrease.
```
name                                   old time/op  new time/op  delta
Scan1_Cockroach-24                      318µs ± 2%   321µs ± 4%    ~     (p=0.529 n=10+10)
Scan1Multinode_Cockroach-24             326µs ± 2%   333µs ± 5%  +2.15%  (p=0.015 n=10+10)
Scan10_Cockroach-24                     357µs ± 1%   357µs ± 3%    ~     (p=0.497 n=9+10)
Scan10Multinode_Cockroach-24            359µs ± 2%   352µs ± 4%    ~     (p=0.075 n=10+10)
Scan100_Cockroach-24                    624µs ± 2%   621µs ± 2%    ~     (p=0.143 n=10+10)
Scan100Multinode_Cockroach-24           616µs ± 2%   606µs ± 2%  -1.65%  (p=0.000 n=9+10)
Scan1000_Cockroach-24                  2.52ms ± 2%  2.51ms ± 1%    ~     (p=0.315 n=10+7)
Scan1000Multinode_Cockroach-24         2.49ms ± 0%  2.51ms ± 1%    ~     (p=0.113 n=10+9)
Scan10000_Cockroach-24                 22.0ms ± 2%  22.3ms ± 2%  +1.34%  (p=0.010 n=10+9)
Scan10000Multinode_Cockroach-24        21.6ms ± 1%  22.4ms ± 3%  +3.45%  (p=0.000 n=8+10)
Scan1000Limit1_Cockroach-24             326µs ± 2%   333µs ± 1%  +2.22%  (p=0.001 n=10+9)
Scan1000Limit1Multinode_Cockroach-24    327µs ± 2%   331µs ± 2%  +1.22%  (p=0.043 n=10+10)
Scan1000Limit10_Cockroach-24            357µs ± 2%   356µs ± 2%    ~     (p=0.684 n=10+10)
Scan1000Limit10Multinode_Cockroach-24   363µs ± 3%   357µs ± 3%    ~     (p=0.089 n=10+10)
Scan1000Limit100_Cockroach-24           618µs ± 1%   625µs ± 3%    ~     (p=0.190 n=9+5)
```

---

Review status: 0 of 26 files reviewed at latest revision, 13 unresolved discussions.

---

*[pkg/sql/scan.go, line 194 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KXl0za31gNIqLhDTEAn:-KXlRQIti_Te4eyxkIkC:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/8717add28055da3027cdd086449ce6bae58032d3/pkg/sql/scan.go#L194)):*
<details><summary><i>Previously, irfansharif (irfan sharif) wrote…</i></summary><blockquote>

/nit `if err := sqlbase.EncDatumRowToDTuple(tuple, encRow, &da); err != nil { ... }` to not reuse `err`/reduce scope. 
</blockquote></details>

Done.

---

*[pkg/sql/sqlbase/encoded_datum.go, line 76 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-KXl49WHBMv-gsPBl6al:-KXlRlXNN-8DUjieUEAj:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/ddaeb83e6872e2fbd42690b280c2449a2cf8c29f/pkg/sql/sqlbase/encoded_datum.go#L76)):*
<details><summary><i>Previously, irfansharif (irfan sharif) wrote…</i></summary><blockquote>

we can do 
```
return EncDatum{
    Type: typ,
    encoding: enc,
    encoded: val,
    Datum: nil,
}
```

</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10550#-:-KXlR7vLs0X5BXTsk-rp:b8y5y1n)*
<!-- Sent from Reviewable.io -->